回复功能异常处理

+++ src/main/java/com/demo/community/dto/CommentDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CommentDTO {
     private Long id;
-    private Long parent_id;
+    private int parent_id;
     private int type;
     private String content;
 }
